sample_llm_rag/requirements.txt
pandas>=1.3.0
python-dotenv>=0.19.0
langchain-core>=0.1.0
langchain-ollama>=0.0.1
langchain-qdrant>=0.0.1
qdrant-client>=1.6.0